Modify 'AuthButton' component
